refactor: extract shared OTLP fixture generation into generators/otlp
Move duplicated OTLP fixture generation code from benches/otlp_io.rs and
bin/otlp_io_profile.rs into a shared module at generators/otlp.rs.

The shared module provides:
- OtlpFixtureProfile struct with all 10 existing Criterion profiles
- 4 new profiles: minimal, single-row, wide-attrs, many-resources
- OtlpFixtureData struct with build_fixture() and all_fixtures()
- ALL_PROFILES constant array with by_name() lookup
- All helper functions: build_request, make_*_attrs, kv_*, severity_*,
  trace_id_for, span_id_for, make_message, complex_body_any_value

Both consumers now import from logfwd_bench::generators::otlp instead of
maintaining their own copies. The Criterion bench version (richer, with
multi-resource/scope support) is used as the canonical source.
