+++ crates/logfwd-bench/src/bin/otlp_io_profile.rs
@@ -8,6 +8,7 @@ use arrow::compute::cast;
 use arrow::datatypes::{DataType, Field, Schema};
 
 use bytes::Bytes;
+use logfwd_bench::generators::otlp::{self, OtlpFixtureProfile};
 use logfwd_bench::{generators, make_otlp_sink};
 use logfwd_io::otlp_receiver::{
     ProjectedOtlpDecoder, decode_protobuf_bytes_to_batch_projected_only_experimental,
@@ -16,10 +17,6 @@ use logfwd_io::otlp_receiver::{
 };
 use logfwd_output::Compression;
 use opentelemetry_proto::tonic::collector::logs::v1::ExportLogsServiceRequest;
-use opentelemetry_proto::tonic::common::v1::any_value::Value;
-use opentelemetry_proto::tonic::common::v1::{AnyValue, InstrumentationScope, KeyValue};
-use opentelemetry_proto::tonic::logs::v1::{LogRecord, ResourceLogs, ScopeLogs};
-use opentelemetry_proto::tonic::resource::v1::Resource;
 use prost::Message as _;
 
 #[cfg(feature = "otlp-profile-alloc")]
@@ -28,39 +25,6 @@ use stats_alloc::{INSTRUMENTED_SYSTEM, Region, StatsAlloc};
 #[cfg(feature = "otlp-profile-alloc")]
 #[global_allocator]
 static GLOBAL: &StatsAlloc<std::alloc::System> = &INSTRUMENTED_SYSTEM;
-
-const NANOS_BASE: u64 = 1_712_509_200_123_456_789;
-
-#[derive(Clone, Copy)]
-struct FixtureProfile {
-    name: &'static str,
-    rows: usize,
-    attrs_per_record: usize,
-    resource_attrs: usize,
-    body_len: usize,
-}
-
-impl FixtureProfile {
-    fn by_name(name: &str) -> Self {
-        match name {
-            "attrs-heavy" => Self {
-                name: "attrs-heavy",
-                rows: 2_000,
-                attrs_per_record: 40,
-                resource_attrs: 8,
-                body_len: 96,
-            },
-            "wide-10k" => Self {
-                name: "wide-10k",
-                rows: 10_000,
-                attrs_per_record: 20,
-                resource_attrs: 6,
-                body_len: 240,
-            },
-            _ => panic!("unknown fixture {name}; expected attrs-heavy or wide-10k"),
-        }
-    }
-}
 
 #[derive(Clone, Copy)]
 enum Mode {
@@ -127,7 +91,7 @@ impl Mode {
 }
 
 struct Cli {
-    case: FixtureProfile,
+    case: OtlpFixtureProfile,
     mode: Option<Mode>,
     iterations: usize,
     flamegraph: Option<PathBuf>,
@@ -187,7 +151,7 @@ fn main() {
 
 impl Cli {
     fn parse() -> Self {
-        let mut case = FixtureProfile::by_name("attrs-heavy");
+        let mut case = OtlpFixtureProfile::by_name("attrs-heavy").expect("attrs-heavy profile");
         let mut mode = None;
         let mut iterations = 1_000usize;
         let mut flamegraph = None;
@@ -197,7 +161,15 @@ impl Cli {
             match arg.as_str() {
                 "--case" => {
                     let value = args.next().expect("--case requires a value");
-                    case = FixtureProfile::by_name(&value);
+                    case = OtlpFixtureProfile::by_name(&value).unwrap_or_else(|| {
+                        panic!(
+                            "unknown fixture {value}; available: {:?}",
+                            otlp::ALL_PROFILES
+                                .iter()
+                                .map(|p| p.name)
+                                .collect::<Vec<_>>()
+                        )
+                    });
                 }
                 "--mode" => {
                     let value = args.next().expect("--mode requires a value");
@@ -230,8 +202,12 @@ impl Cli {
 
 fn print_usage() {
     eprintln!(
-        "Usage: otlp_io_profile [--case attrs-heavy|wide-10k] [--mode MODE] [--iterations N] [--flamegraph PATH]"
+        "Usage: otlp_io_profile [--case PROFILE] [--mode MODE] [--iterations N] [--flamegraph PATH]"
     );
+    eprintln!("Profiles:");
+    for profile in &otlp::ALL_PROFILES {
+        eprintln!("  {}", profile.name);
+    }
     eprintln!("Modes:");
     for mode in Mode::ALL {
         eprintln!("  {}", mode.name());
@@ -418,28 +394,34 @@ fn print_result(result: &ProfileResult) {
     );
 }
 
-fn build_fixture(profile: FixtureProfile) -> FixtureData {
-    let payload = build_request(profile).encode_to_vec();
-    let payload_bytes = Bytes::from(payload.clone());
-    let batch = decode_protobuf_to_batch_prost_reference(&payload).expect("fixture decodes");
-    assert_eq!(batch.num_rows(), profile.rows);
-    let production = decode_protobuf_to_batch(&payload).expect("production fixture decodes");
+fn build_fixture(profile: OtlpFixtureProfile) -> FixtureData {
+    let otlp_data = otlp::build_fixture(profile);
+    let batch =
+        decode_protobuf_to_batch_prost_reference(&otlp_data.payload).expect("fixture decodes");
+    assert_eq!(batch.num_rows(), profile.total_rows());
+    let production =
+        decode_protobuf_to_batch(&otlp_data.payload).expect("production fixture decodes");
     assert_batch_matches(&batch, &production, profile.name);
     assert_encode_paths_match(&production, profile.name);
-    let projected = decode_protobuf_to_batch_projected_detached_experimental(&payload)
-        .expect("projected fixture decodes");
-    assert_batch_matches(&batch, &projected, profile.name);
-    assert_encode_paths_match(&projected, profile.name);
-    let view = decode_protobuf_bytes_to_batch_projected_only_experimental(payload_bytes.clone())
+    if !profile.has_complex_any {
+        let projected =
+            decode_protobuf_to_batch_projected_detached_experimental(&otlp_data.payload)
+                .expect("projected fixture decodes");
+        assert_batch_matches(&batch, &projected, profile.name);
+        assert_encode_paths_match(&projected, profile.name);
+        let view = decode_protobuf_bytes_to_batch_projected_only_experimental(
+            otlp_data.payload_bytes.clone(),
+        )
         .expect("view fixture decodes");
-    let detached_view = logfwd_arrow::materialize::detach(&view);
-    assert_batch_matches(&batch, &detached_view, profile.name);
-    assert_encode_paths_match(&view, profile.name);
+        let detached_view = logfwd_arrow::materialize::detach(&view);
+        assert_batch_matches(&batch, &detached_view, profile.name);
+        assert_encode_paths_match(&view, profile.name);
+    }
 
     FixtureData {
-        payload,
-        payload_bytes,
-        rows: profile.rows,
+        payload: otlp_data.payload,
+        payload_bytes: otlp_data.payload_bytes,
+        rows: profile.total_rows(),
     }
 }
 
@@ -517,197 +499,4 @@ fn assert_batch_matches(
             "profile fixture column {idx} must match prost batch for {fixture_name}"
         );
     }
-}
-
-fn build_request(profile: FixtureProfile) -> ExportLogsServiceRequest {
-    let mut records = Vec::with_capacity(profile.rows);
-    for row in 0..profile.rows {
-        records.push(LogRecord {
-            time_unix_nano: NANOS_BASE + row as u64,
-            observed_time_unix_nano: NANOS_BASE + row as u64 + 123,
-            severity_text: severity_text(row).to_string(),
-            severity_number: severity_number(row),
-            body: Some(AnyValue {
-                value: Some(Value::StringValue(make_message(row, profile.body_len))),
-            }),
-            trace_id: trace_id_for(row),
-            span_id: span_id_for(row),
-            flags: 1,
-            attributes: make_record_attrs(profile, row),
-            ..Default::default()
-        });
-    }
-
-    ExportLogsServiceRequest {
-        resource_logs: vec![ResourceLogs {
-            resource: Some(Resource {
-                attributes: make_resource_attrs(profile),
-                ..Default::default()
-            }),
-            scope_logs: vec![ScopeLogs {
-                scope: Some(InstrumentationScope {
-                    name: "bench-scope-0-0".to_string(),
-                    version: "2026.04".to_string(),
-                    attributes: vec![kv_string("scope.env", "bench")],
-                    dropped_attributes_count: 0,
-                }),
-                log_records: records,
-                schema_url: String::new(),
-            }],
-            schema_url: String::new(),
-        }],
-    }
-}
-
-fn make_resource_attrs(profile: FixtureProfile) -> Vec<KeyValue> {
-    let mut attrs = Vec::with_capacity(profile.resource_attrs.max(4));
-    attrs.push(kv_string("service.name", "checkout"));
-    attrs.push(kv_string("service.namespace", "payments"));
-    attrs.push(kv_string("deployment.environment", "benchmark"));
-    attrs.push(kv_string("host.name", "bench-host-0"));
-    for idx in 0..profile.resource_attrs.saturating_sub(4) {
-        attrs.push(kv_string(
-            format!("resource.extra.{idx}"),
-            format!("resource-value-{idx}"),
-        ));
-    }
-    attrs
-}
-
-fn make_record_attrs(profile: FixtureProfile, row: usize) -> Vec<KeyValue> {
-    let mut attrs = Vec::with_capacity(profile.attrs_per_record);
-    attrs.push(kv_string("http.method", method(row)));
-    attrs.push(kv_string(
-        "http.route",
-        format!("/api/v1/item/{}", row % 128),
-    ));
-    attrs.push(kv_int("http.status_code", status_code(row)));
-    attrs.push(kv_bool("cache.hit", row.is_multiple_of(3)));
-    for idx in 0..profile.attrs_per_record.saturating_sub(4) {
-        match idx % 5 {
-            0 => attrs.push(kv_string(
-                format!("attr.string.{idx}"),
-                format!("value-{row}-{idx}"),
-            )),
-            1 => attrs.push(kv_int(format!("attr.int.{idx}"), (row * (idx + 1)) as i64)),
-            2 => attrs.push(kv_bool(
-                format!("attr.bool.{idx}"),
-                (row + idx).is_multiple_of(2),
-            )),
-            3 => attrs.push(kv_double(
-                format!("attr.double.{idx}"),
-                row as f64 * 0.25 + idx as f64,
-            )),
-            _ => attrs.push(kv_bytes(
-                format!("attr.bytes.{idx}"),
-                vec![row as u8, idx as u8, (row >> 8) as u8, 0xab],
-            )),
-        }
-    }
-    attrs
-}
-
-fn kv_string(key: impl Into<String>, value: impl Into<String>) -> KeyValue {
-    KeyValue {
-        key: key.into(),
-        value: Some(AnyValue {
-            value: Some(Value::StringValue(value.into())),
-        }),
-    }
-}
-
-fn kv_int(key: impl Into<String>, value: i64) -> KeyValue {
-    KeyValue {
-        key: key.into(),
-        value: Some(AnyValue {
-            value: Some(Value::IntValue(value)),
-        }),
-    }
-}
-
-fn kv_bool(key: impl Into<String>, value: bool) -> KeyValue {
-    KeyValue {
-        key: key.into(),
-        value: Some(AnyValue {
-            value: Some(Value::BoolValue(value)),
-        }),
-    }
-}
-
-fn kv_double(key: impl Into<String>, value: f64) -> KeyValue {
-    KeyValue {
-        key: key.into(),
-        value: Some(AnyValue {
-            value: Some(Value::DoubleValue(value)),
-        }),
-    }
-}
-
-fn kv_bytes(key: impl Into<String>, value: Vec<u8>) -> KeyValue {
-    KeyValue {
-        key: key.into(),
-        value: Some(AnyValue {
-            value: Some(Value::BytesValue(value)),
-        }),
-    }
-}
-
-fn severity_text(row: usize) -> &'static str {
-    match row % 5 {
-        0 => "TRACE",
-        1 => "DEBUG",
-        2 => "INFO",
-        3 => "WARN",
-        _ => "ERROR",
-    }
-}
-
-fn severity_number(row: usize) -> i32 {
-    match row % 5 {
-        0 => 1,
-        1 => 5,
-        2 => 9,
-        3 => 13,
-        _ => 17,
-    }
-}
-
-fn method(row: usize) -> &'static str {
-    match row % 4 {
-        0 => "GET",
-        1 => "POST",
-        2 => "PUT",
-        _ => "DELETE",
-    }
-}
-
-fn status_code(row: usize) -> i64 {
-    match row % 20 {
-        0 => 500,
-        1 | 2 => 404,
-        _ => 200,
-    }
-}
-
-fn make_message(row: usize, len: usize) -> String {
-    let mut message = format!("benchmark row={row} subsystem=otlp_io ");
-    while message.len() < len {
-        message.push_str("payload ");
-    }
-    message.truncate(len);
-    message
-}
-
-fn trace_id_for(row: usize) -> Vec<u8> {
-    let mut bytes = [0u8; 16];
-    bytes[..8].copy_from_slice(&(row as u64).to_be_bytes());
-    bytes[8..].copy_from_slice(&((row as u64).wrapping_mul(0x9e37)).to_be_bytes());
-    bytes.to_vec()
-}
-
-fn span_id_for(row: usize) -> Vec<u8> {
-    (row as u64)
-        .wrapping_mul(0x517c_c1b7_2722_0a95)
-        .to_be_bytes()
-        .to_vec()
 }
